test: move perps migration CV coverage into component tests
